% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_adults}
\alias{df_adults}
\title{Example of WHO Model Disability Survey data for adults}
\format{A tibble with 2500 rows and 90 variables:
\describe{
  \item{HHID}{household number}
  \item{strata}{survey strata}
  \item{PSU}{PSU id}
  \item{weight}{survey weights}
  \item{sex}{sex, Male or Female}
  \item{age}{age in years}
  \item{age_cat}{age catgory, one of "18-24", "25-39", "40-64", or "64-100"}
  \item{work_cat}{whether respondent is currently working, binary}
  \item{edu_cat}{highest level of education attained}
  \item{F....}{survey item from "Functioning" section}
  \item{C....}{survey item from "Capacity" section}
  \item{EF....}{survey item from "Environmental factors" section}
  \item{disability_score}{metric scale of performance, from Rasch analysis}
  \item{disability_cat}{performance level, as definted by WHO cutoffs of disability_score}
}}
\usage{
df_adults
}
\description{
An dummy data set of data from the WHO Model Disability Survey.
}
\keyword{datasets}
