% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_children}
\alias{df_children}
\title{Example of WHO Model Disability Survey data for children}
\format{A tibble with 2500 rows and 42 variables:
\describe{
  \item{HHID}{household number}
  \item{strata}{survey strata}
  \item{PSU}{PSU id}
  \item{weight}{survey weights}
  \item{sex}{sex, Male or Female}
  \item{age}{age in years}
  \item{age_cat}{age catgory, one of "2-4", "5-9",  or"10-17"}
  \item{child....}{survey item from MDS Children questionnaire"}
}}
\usage{
df_children
}
\description{
An dummy data set of data from the WHO Model Disability Survey.
}
\keyword{datasets}
