% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_poppyramid.R
\name{fig_poppyramid}
\alias{fig_poppyramid}
\title{Print a population pyramid}
\usage{
fig_poppyramid(df, var_age, var_sex)
}
\arguments{
\item{df}{a tibble of household survey data where each row is a household member}

\item{var_age}{a string (length 1) of the name of the age column}

\item{var_sex}{a string (length 1) of the name of the sex column}
}
\value{
A population pyramid figure
}
\description{
Print a population pyramid
}
\examples{
fig_poppyramid(mdstest, "edad", "sex")
}
