% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_installation.R
\name{helper_installation}
\alias{helper_installation}
\title{Check installation of whomds is the most updated}
\usage{
helper_installation()
}
\value{
Prints a message stating whether or not installed package is same as most updated version from Github
}
\description{
Check installation of whomds is the most updated
}
\examples{
whomds:::helper_installation()
}
