% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_palette.R
\name{helper_palette}
\alias{helper_palette}
\title{Color palette for the MDS}
\usage{
helper_palette(n, h = c(-100, 100), c. = c(60, 100), l = c(15, 95),
  power = c(2, 0.9), fixup = TRUE, gamma = NULL, alpha = 1, ...)
}
\value{
\code{n} hex codes from the color palette
}
\description{
Color palette for the MDS
}
\details{
Generated with \code{colorspace::choose_palette()}
}
\seealso{
Other helper functions: \code{\link{helper_varslist}}
}
