% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_varslist.R
\name{helper_varslist}
\alias{helper_varslist}
\title{Convert list to vector with unique elements}
\usage{
helper_varslist(vars_list)
}
\arguments{
\item{vars_list}{a list of character vectors}
}
\value{
a character vector of all unique items from \code{vars_list}
}
\description{
Convert list to vector with unique elements
}
\seealso{
Other helper functions: 
\code{\link{helper_checkrow}()},
\code{\link{helper_palette}()}
}
\concept{helper functions}
