% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_DIF.R
\name{rasch_DIF}
\alias{rasch_DIF}
\title{Perform analysis of Differential Item Functioning (DIF) for Rasch Model}
\usage{
rasch_DIF(df, vars_metric, vars_DIF, residuals_PCM, split_strategy = NULL,
  print_results = TRUE, path_output = NULL, breaks = 6)
}
\arguments{
\item{df}{a tibble of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{vars_DIF}{a string with the column names to use for analyzing differential item functioning (DIF)}

\item{residuals_PCM}{a matrix giving the residuals of the person parameters from the Rasch Model. Row names are the numbers of the people and the columns are for each variable.}

\item{split_strategy}{a named list giving the strategy to take for spliting variables by categories, passed to \code{rasch_split()}. One element of the list per variable to split by. Each element of the list must be a character vector of column names to split. The names of the list are the variables to split each group of variables by. Default is NULL, to indicate there was no splitting.}

\item{print_results}{a logical value indicating whether or not to print various files displaying results from the Rasch Model. Default is TRUE, to print the files.}

\item{path_output}{a string with the path to the output folder. Default is NULL.}

\item{breaks}{a numeric value giving the number if class intervals. Default is 6.}
}
\value{
a list with results from the DIF analysis
}
\description{
Perform analysis of Differential Item Functioning (DIF) for Rasch Model
}
\details{
Currently the calculation of the class intervals is quite slow. Reducing the number of breaks can improve speed.
}
