% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_drop.R
\name{rasch_drop}
\alias{rasch_drop}
\title{Drop items from a Rasch Analysis}
\usage{
rasch_drop(vars_metric, drop_vars, max_values)
}
\arguments{
\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{drop_vars}{a character vector of items to drop from the Rasch Analysis}

\item{max_values}{a tibble with two columns, \code{var} equivalent to \code{vars_metric} and \code{max_val} with their corresponding maximum possible values}
}
\value{
a named list with the new \code{vars_metric} and new \code{max_values} after dropping the desired variables
}
\description{
Drop items from a Rasch Analysis
}
\details{
Dropping variables might be desirable if one finds that particular items are causing a lot of problems for the fit of a Rasch Model.
}
\seealso{
Other rasch functions: \code{\link{rasch_DIF}},
  \code{\link{rasch_factor}},
  \code{\link{rasch_mds_children}},
  \code{\link{rasch_mds}}, \code{\link{rasch_model}},
  \code{\link{rasch_rawscore}}, \code{\link{rasch_rescale}}
}
