% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_factor.R
\name{rasch_factor}
\alias{rasch_factor}
\title{Calculate a factor analysis for a Rasch Model}
\usage{
rasch_factor(df, vars_metric, print_results = TRUE)
}
\arguments{
\item{df}{a tibble of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{print_results}{a logical vector indicating whether to print the results of the model to the \code{model_name} directory}
}
\value{
a list with results from the factor analysis for a Rasch Model
}
\description{
Calculate a factor analysis for a Rasch Model
}
