% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_model.R
\name{rasch_model}
\alias{rasch_model}
\title{Run the Rasch Model and print diagnositic results}
\usage{
rasch_model(df, vars_metric, print_results = TRUE, LIDcutoff = 0.2)
}
\arguments{
\item{df}{a tibble of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{print_results}{a logical vector indicating whether to print the results of the model to the \code{model_name} directory}

\item{LIDcutoff}{a numeric value between 0 and 1 indicating the cut-off for significant local item dependence}
}
\value{
a list \code{model_result} with results from the Rasch Model
}
\description{
Run the Rasch Model and print diagnositic results
}
