% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_recode.R
\name{rasch_recode}
\alias{rasch_recode}
\title{Recode survey items for use in Rasch Analysis}
\usage{
rasch_recode(df, vars_metric, recode_strategy, max_values)
}
\arguments{
\item{df}{a data frame of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{recode_strategy}{a named list giving the strategy to take for recoding variables, passed to \code{rasch_recode()}. One element of the list per recode strategy. Each element of the list is a numeric vector giving the new values to map the variables to. The names of the list are the groups of column names to use for each recoding strategy, separated only by ",".}

\item{max_values}{a tibble with two columns, \code{var} equivalent to \code{vars_metric} and \code{max_val} with their corresponding maximum possible values}
}
\value{
a named list with the new \code{df} and new \code{max_values} after recoding the desired variables
}
\description{
Recode survey items for use in Rasch Analysis
}
