% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_testlet.R
\name{rasch_testlet}
\alias{rasch_testlet}
\title{Create testlets of survey items for a Rasch Model}
\usage{
rasch_testlet(df, vars_metric, testlet_strategy, max_values, resp_opts)
}
\arguments{
\item{df}{a tibble of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{testlet_strategy}{a list giving the strategy to take for creating testlets, passed to \code{rasch_testlet()}. One element of the list per testlet to create. Each element of the list must be a character vector of column names to use for the testlet. Optionally, name the element of the list to give the name of the new testlet. Otherwise, the new testlet will be the original column names separated by "_".}

\item{max_values}{a tibble with two columns, \code{var} equivalent to \code{vars_metric} and \code{max_val} with their corresponding maximum possible values}

\item{resp_opts}{a numeric vector of possible response options for \code{vars_metric}.}
}
\value{
a named list with the new \code{df}, new \code{vars_metric}, new \code{testlet_strategy} and new \code{max_values} after creating desired testlets
}
\description{
Create testlets of survey items for a Rasch Model
}
