% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_weightedpct.R
\name{table_weightedpct}
\alias{table_weightedpct}
\title{Calculate table of percentages of response distribution for survey items, survey weighted, disaggregated}
\usage{
table_weightedpct(df, vars_ids, vars_strata, vars_weights, formula_vars,
  formula_vars_levels = 0:1, by_vars, spread_key = NULL,
  spread_value = "prop", arrange_vars = NULL, willfilter = NULL, ...)
}
\arguments{
\item{df}{a tibble of survey responses, one row per respondent}

\item{vars_ids}{a character vector of cluster ids, passed to a \code{survey::svydesign} object}

\item{vars_strata}{a character vector of strata ids, passed to a \code{survey::svydesign} object}

\item{vars_weights}{a character vector of survey weight ids, passed to a \code{survey::svydesign} object}

\item{formula_vars}{a character vector of variables to calculate the percentages of each level for}

\item{formula_vars_levels}{a vector of the levels of the the \code{formula_vars}}

\item{by_vars}{a character vector of variables to disaggregate results by}

\item{spread_key}{a string with variable name to pass to \code{key} argument of \code{dplyr::spread()}. Default is NULL.}

\item{spread_value}{a string with variable name to pass to \code{value} argument of \code{dplyr::spread()}. Default is "prop" (the columm of percentages created within the function)}

\item{arrange_vars}{a character vector with variables to pass to \code{dplyr::arrange()}. Default is NULL.}

\item{willfilter}{a logical variable that tells the function whether or not to filter or transmute the data. Leave as default NULL to not filter or transmute. Set as TRUE to filter and FALSE to transmute. See Details.}

\item{...}{captures expressions to pass to \code{dplyr::filter()} or \code{dplyr::transmute()}, depending on the value of argument \code{willfilter}. See Details.}
}
\value{
a tibble of weighted response percentages
}
\description{
Calculate table of percentages of response distribution for survey items, survey weighted, disaggregated
}
\details{
ADD SOME STUFF ABOUT THE FILTERING
}
\examples{
table_weightedpct(mdstest, 
    vars_ids = c("VARUNIT_N","enc_idr"),
    vars_strata = "VARSTRAT_N",
    vars_weights = "Factor_Persona",
    formula_vars = paste0("fa",1:10),
    formula_vars_levels = 1:5,
    by_vars = "sex")
}
