% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomds.R
\docType{package}
\name{whomds}
\alias{whomds}
\alias{whomds-package}
\title{whomds: A package for calculating results from WHO Model Disability Survey}
\description{
The whomds package provides two categories of important functions:
table functions and graph functions
}
\section{Table functions}{

The table functions output different fit for purpose tables for reporting
results from the WHO Model Disability Survey (MDS)
}

\section{Graph functions}{

The graph functions output different graphs for reporting
results from the WHO Model Disability Survey (MDS)
}

