% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_poppyramid.R
\name{fig_poppyramid}
\alias{fig_poppyramid}
\title{Print a population pyramid}
\usage{
fig_poppyramid(df, var_age, var_sex)
}
\arguments{
\item{df}{a data frame of individual survey data, where each row is an individual}

\item{var_age}{a string (length 1) of the name of the age column}

\item{var_sex}{a string (length 1) of the name of the sex column}
}
\value{
A population pyramid figure
}
\description{
Print a population pyramid
}
\details{
A population pyramid gives an image of the age and sex distribution of a population.

The function divides the \code{var_age} variable into 5-year intervals.
}
\examples{
fig_poppyramid(chile_adults, "edad", "sex")
}
\seealso{
Other figure functions: \code{\link{fig_LID}},
  \code{\link{fig_dist}}
}
\concept{figure functions}
