% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_table.R
\name{prev_table}
\alias{prev_table}
\title{Calculate a prevalence table}
\usage{
prev_table(data, id, weight = NULL, strata.svy = NULL, prev_vars,
  lev = 1:5, which.prev = lev, stratum = NULL, disagg = NULL,
  stack = FALSE)
}
\arguments{
\item{data}{a data.table of the survey data}

\item{id}{a character vector (length 1) with the name of the column with the individual IDs}

\item{weight}{a character vector (length 1) with the name of the column with the survey weights, default is NULL}

\item{strata.svy}{a character vector (length 1) with the name of the column with the survey strata, default is NULL}

\item{prev_vars}{a character vector with the names of the survey items for which prevalences should be calculated}

\item{lev}{a numeric vector giving the response options of prev_vars, default is 1:5}

\item{which.prev}{a numeric vector giving the index of the response options that you want prevalences printed for, default is equal to lev. For example if lev=0:1 and you only want prevalences for response option 1 (the second response option), then which.prev=2}

\item{stratum}{a character vector (length 1) with the name of the column to stratify by, default is NULL}

\item{disagg}{a character vector (length 1) with the name of the column by which to perform a second level of disaggregation (e.g. sex), default is NULL}

\item{stack}{a logical vector (length 1) indicating whether the prevalence table should be stacked, default is FALSE. This is most useful when prev_vars is length 1}
}
\value{
A table with the prevalences for each response option for the given prev_vars, stratified by stratum and disaggregated by disagg
}
\description{
Calculate a prevalence table
}
