% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_drop.R
\name{rasch_drop}
\alias{rasch_drop}
\title{Drop items from a Rasch Analysis}
\usage{
rasch_drop(vars_metric, drop_vars, max_values)
}
\arguments{
\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{drop_vars}{a character vector of items to drop from the Rasch Analysis}

\item{max_values}{a tibble with two columns, \code{var} equivalent to \code{vars_metric} and \code{max_val} with their corresponding maximum possible values}
}
\value{
a named list with the new \code{vars_metric} and new \code{max_values} after dropping the desired variables
}
\description{
Drop items from a Rasch Analysis
}
