% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_split.R
\name{rasch_split}
\alias{rasch_split}
\title{Split survey items by categories for a Rasch Model}
\usage{
rasch_split(df, vars_metric, split_strategy, max_values)
}
\arguments{
\item{df}{a tibble of individual survey data, where each row is an individual}

\item{vars_metric}{a character vector of items to use in the Rasch Analysis}

\item{split_strategy}{a named list giving the strategy to take for spliting variables by categories, passed to \code{rasch_split()}. One element of the list per variable to split by. Each element of the list must be a character vector of column names to split. The names of the list are the variables to split each group of variables by.}

\item{max_values}{a tibble with two columns, \code{var} equivalent to \code{vars_metric} and \code{max_val} with their corresponding maximum possible values}
}
\value{
a named list with the new \code{df}, new \code{vars_metric} and new \code{max_values} after splitting the desired variables
}
\description{
Split survey items by categories for a Rasch Model
}
\details{
If significant differential item functioning (DIF) is observed, it may be desirable to split variables based on the charactersitic for which DIF is observed. For example, if men and women have significantly different patterns of responses to items, then it may be desirable to split items by sex. This function performs that variable splitting.
}
