% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_unweightedpctn.R
\name{table_unweightedpctn}
\alias{table_unweightedpctn}
\title{Compute unweighted percent and N for multiple variables, disaggregated}
\usage{
table_unweightedpctn(df, vars_demo, group_by_var = NULL,
  spread_by_group_by_var = FALSE)
}
\arguments{
\item{df}{a tibble of individual data where each row is an individual}

\item{vars_demo}{a character vector of names of variables to calculate percent and N for}

\item{group_by_var}{a string (length 1) with the name of the variable from \code{df} to disaggregate by}

\item{spread_by_group_by_var}{logical determining whether to pass \code{group_by_var} to \code{dplyr::spread()} to give a wide-format tab. Default is FALSE.}
}
\value{
A tibble with percent and N for each level of each variable in \code{vars_demo}
}
\description{
Compute unweighted percent and N for multiple variables, disaggregated
}
\examples{
table_unweightedpctn(mdstest, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"))
table_unweightedpctn(mdstest, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"), 
group_by_var = "performance_cat")
table_unweightedpctn(mdstest, vars_demo = c("sex", "age_cat", "work_cat", "edu_cat"), 
group_by_var = "performance_cat", spread_by_group_by_var = TRUE)
}
