% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_weightedpct.R
\name{table_weightedpct}
\alias{table_weightedpct}
\title{Calculate table of percentages of response distribution for survey items, survey weighted, disaggregated}
\usage{
table_weightedpct(df, vars_ids, vars_strata, vars_weights, formula_vars, ...,
  formula_vars_levels = 0:1, by_vars = NULL, spread_key = NULL,
  spread_value = "prop", arrange_vars = NULL, willfilter = NULL)
}
\arguments{
\item{df}{a data frame of survey responses, one row per respondent}

\item{vars_ids}{a character vector of cluster ids, passed to a \code{survey::svydesign} object}

\item{vars_strata}{a character vector of strata ids, passed to a \code{survey::svydesign} object}

\item{vars_weights}{a character vector of survey weight ids, passed to a \code{survey::svydesign} object}

\item{formula_vars}{a character vector of variables to calculate the percentages of each level for}

\item{...}{captures expressions to pass to \code{dplyr::filter()} or \code{dplyr::transmute()}, depending on the value of argument \code{willfilter}. See Details.}

\item{formula_vars_levels}{a vector of the levels of the the \code{formula_vars}}

\item{by_vars}{a character vector of variables to disaggregate results by. Default is \code{NULL} for no disaggregation.}

\item{spread_key}{a string with variable name to pass to \code{key} argument of \code{dplyr::spread()}. Default is \code{NULL}.}

\item{spread_value}{a string with variable name to pass to \code{value} argument of \code{dplyr::spread()}. Default is "prop" (the columm of percentages created within the function)}

\item{arrange_vars}{a character vector with variables to pass to \code{dplyr::arrange()}. Default is NULL.}

\item{willfilter}{a logical variable that tells the function whether or not to filter or transmute the data. Leave as default \code{NULL} to not filter or transmute. Set as \code{TRUE} to filter and \code{FALSE} to transmute. See Details.}
}
\value{
a tibble of weighted response percentages
}
\description{
Calculate table of percentages of response distribution for survey items, survey weighted, disaggregated
}
\details{
If \code{willfilter} is NULL, the table is not filtered or transmuted. If \code{willfilter} is TRUE, the table is filtered before it is spread or arranged. If \code{willfilter} is FALSE, the table is transmuted after the spread and/or arrange. "..." captures the non-standard evaluation expressions (NSE) to pass to \code{dplyr::filter} or \code{dplyr::transmute()}.
}
\examples{
table_weightedpct(mdstest, 
    vars_ids = c("VARUNIT_N","enc_id"),
    vars_strata = "VARSTRAT_N",
    vars_weights = "Factor_Persona",
    formula_vars = paste0("fa",1:10),
    formula_vars_levels = 1:5,
    by_vars = "sex")
}
\seealso{
See \code{vignette("programming", package = "dplyr")} for more about non-standard evaluation (NSE)

Other table functions: \code{\link{table_basicstats}},
  \code{\link{table_unweightedpctn}}
}
