% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomds.R
\docType{package}
\name{whomds}
\alias{whomds}
\title{whomds: A package for calculating results from WHO Model Disability Survey}
\description{
The whomds package provides three categories of important functions:
table functions, figure functions, and Rasch Analysis functions
}
\section{Table functions}{

The table functions output different fit for purpose tables for reporting
results from the WHO Model Disability Survey (MDS). They begin with \code{table_*()}
}

\section{Figure functions}{

The figures functions output different figures for reporting
results from the WHO Model Disability Survey (MDS). They begin with \code{fig_*()}
}

\section{Rasch Analysis functions}{

These functions are used to complete an iteration of Rasch Analysis for WHO Model Disability Survey (MDS). They begin with \code{rasch_*}
}

\references{
WHO Model Disability Survey: \url{https://www.who.int/health-topics/disability}
}
